requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==21.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.10.3
jinja2==3.1.4
jsonschema==4.23.0
python-dotenv==1.0.1
google-generativeai==0.8.3
google-api-python-client==2.114.0
google-auth==2.27.0
apscheduler==3.10.4
python-dateutil==2.9.0
requests==2.31.0
